+++ xc3_model/src/skinning.rs
@@ -105,7 +105,7 @@ impl WeightGroups {
     pub fn weights_start_index(
         &self,
         flags2: u32,
-        lod: u16,
+        lod: u8,
         unk_type: xc3_lib::mxmd::RenderPassType,
     ) -> usize {
         match self {
@@ -128,14 +128,14 @@ impl WeightGroups {
 fn weight_group_index(
     weight_lods: &[WeightLod],
     skin_flags: u32,
-    lod: u16,
+    lod: u8,
     unk_type: RenderPassType,
 ) -> usize {
     if !weight_lods.is_empty() {
         // TODO: Should this check skin flags?
         // TODO: Is lod actually some sort of flags?
         // TODO: Return none if skin_flags == 64?
-        let lod_index = (lod & 0xff).saturating_sub(1) as usize;
+        let lod_index = lod.saturating_sub(1) as usize;
         // TODO: More mesh lods than weight lods for models with multiple lod groups?
         let weight_lod = &weight_lods[lod_index % weight_lods.len()];
 
@@ -319,6 +319,7 @@ impl SkinWeights {
                     }
                 }
             } else {
+                // TODO: This can result in bone names not working?
                 error!("Influence {:?} not found in skeleton.", influence.bone_name);
             }
         }