+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Fixed an issue loading mipmaps for some Xenoblade X textures.
 * Fixed an issue where alpha meshes for Xenoblade X models did not render.
+* Fixed an issue where some models did not load the correct meshes for the base level of detail (LOD).
 
 ## 0.9.0 - 2024-05-17
 ### Added